# artifacts/api-server/src/routes/storage.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { Readable } from "stream";
import { eq } from "drizzle-orm";
import { db, evidenceTable, complaintsTable } from "@workspace/db";
import {
  RequestUploadUrlBody,
  RequestUploadUrlResponse,
} from "@workspace/api-zod";
import { ObjectStorageService, ObjectNotFoundError } from "../lib/objectStorage";
import { ObjectPermission } from "../lib/objectAcl";
import { canAccessComplaint } from "../middlewares/rbac";
import { logAudit } from "../lib/audit";

const router: IRouter = Router();
const objectStorageService = new ObjectStorageService();

/**
 * POST /storage/uploads/request-url
 *
 * Request a presigned URL for file upload.
 * The client sends JSON metadata (name, size, contentType) — NOT the file.
 * Then uploads the file directly to the returned presigned URL.
 */
router.post("/storage/uploads/request-url", async (req: Request, res: Response) => {
  if (!req.localUser) {
    res.status(401).json({ error: "Authentication required to upload files" });
    return;
  }

  const parsed = RequestUploadUrlBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Missing or invalid required fields" });
    return;
  }

  try {
    const { name, size, contentType } = parsed.data;

    const uploadURL = await objectStorageService.getObjectEntityUploadURL();
    const objectPath = objectStorageService.normalizeObjectEntityPath(uploadURL);

    res.json(
      RequestUploadUrlResponse.parse({
        uploadURL,
        objectPath,
        metadata: { name, size, contentType },
      }),
    );
  } catch (error) {
    req.log.error({ err: error }, "Error generating upload URL");
    res.status(500).json({ error: "Failed to generate upload URL" });
  }
});

/**
 * GET /storage/public-objects/*
 *
 * Serve public assets from PUBLIC_OBJECT_SEARCH_PATHS.
 * These are unconditionally public — no authentication or ACL checks.
 * IMPORTANT: Always provide this endpoint when object storage is set up.
 */
router.get("/storage/public-objects/*filePath", async (req: Request, res: Response) => {
  try {
    const raw = req.params.filePath;
    const filePath = Array.isArray(raw) ? raw.join("/") : raw;
    const file = await objectStorageService.searchPublicObject(filePath);
    if (!file) {
      res.status(404).json({ error: "File not found" });
      return;
    }

    const response = await objectStorageService.downloadObject(file);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    req.log.error({ err: error }, "Error serving public object");
    res.status(500).json({ error: "Failed to serve public object" });
  }
});

/**
 * GET /storage/objects/*
 *
 * Serve private object entities from PRIVATE_OBJECT_DIR.
 * Requires authentication. ACL is enforced per-object: the requesting user
 * must be the owner stored in the object's ACL policy metadata.
 * Fails closed — objects with no ACL policy are denied.
 */
router.get("/storage/objects/*path", async (req: Request, res: Response) => {
  try {
    if (!req.localUser) {
      res.status(401).json({ error: "Authentication required to access private files" });
      return;
    }

    const raw = req.params.path;
    const wildcardPath = Array.isArray(raw) ? raw.join("/") : raw;
    const objectPath = `/objects/${wildcardPath}`;
    const objectFile = await objectStorageService.getObjectEntityFile(objectPath);

    // Complaint evidence files are authorized by complaint context (owner,
    // assigned investigator, jurisdiction-authorized officer roles) rather
    // than uploader-only object ACL.
    const evidenceRows = await db
      .select({
        evidenceId: evidenceTable.id,
        complaintId: complaintsTable.id,
        complaintUserId: complaintsTable.userId,
        departmentId: complaintsTable.departmentId,
        districtId: complaintsTable.districtId,
        assignedOfficerId: complaintsTable.assignedOfficerId,
      })
      .from(evidenceTable)
      .innerJoin(complaintsTable, eq(evidenceTable.complaintId, complaintsTable.id))
      .where(eq(evidenceTable.fileUrl, objectPath))
      .limit(1);

    if (evidenceRows[0]) {
      const ev = evidenceRows[0];
      const user = req.localUser;
      if (!user) {
        res.status(403).json({ error: "Access denied" });
        return;
      }
      const allowed =
        user.role !== "citizen"
          ? canAccessComplaint(user, ev)
          : ev.complaintUserId !== null && ev.complaintUserId === user.id;
      if (!allowed) {
        res.status(403).json({ error: "Access denied" });
        return;
      }
      // Audit the actual file download (not just metadata listing)
      await logAudit({
        req,
        userId: user.id,
        action: "evidence_download",
        entityType: "evidence",
        entityId: ev.evidenceId,
        details: { complaintId: ev.complaintId, objectPath },
      });
    } else {
      // Non-evidence objects: fall back to per-object ACL (fails closed)
      const canAccess = await objectStorageService.canAccessObjectEntity({
        userId: String(req.localUser.id),
        objectFile,
        requestedPermission: ObjectPermission.READ,
      });
      if (!canAccess) {
        res.status(403).json({ error: "Access denied" });
        return;
      }
    }

    const response = await objectStorageService.downloadObject(objectFile);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    if (error instanceof ObjectNotFoundError) {
      req.log.warn({ err: error }, "Object not found");
      res.status(404).json({ error: "Object not found" });
      return;
    }
    req.log.error({ err: error }, "Error serving object");
    res.status(500).json({ error: "Failed to serve object" });
  }
});

export default router;