# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CorruptionFreeTN API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /**
     * Email address or mobile number
     * @minLength 1
     */
  identifier: string;
  /** @minLength 1 */
  password: string;
}