# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CorruptionFreeTN API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @maxLength 120 */
  name?: string;
  email?: string;
  phone?: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}