# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CorruptionFreeTN API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  districtId?: number | null;
}